% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy_classes.R
\docType{class}
\name{CPPI-class}
\alias{CPPI-class}
\title{An S4 class for the CPPI hedging strategy}
\description{
An S4 class for the CPPI hedging strategy
}
\section{Slots}{

\describe{
\item{\code{RiskFactor}}{The risk factor (cushion) used in the CPPI model}
}}

