% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CppiStrat.R
\name{CppiStrat}
\alias{CppiStrat}
\title{Constant Proportion Portfolio Insurance (CPPI)}
\usage{
CppiStrat(Q, P, TP, RF = 10, tcost = 0, min = Q, int = TRUE)
}
\arguments{
\item{Q}{numeric value for quantity to be hedged}

\item{P}{numeric futures price vector}

\item{TP}{numeric target price to protect}

\item{RF}{numeric risk factor}

\item{tcost}{numeric transaction costs pr unit}

\item{min}{numeric minimum quantity to be hedged at delivery}

\item{int}{TRUE/ FALSE integer restriction on tradable volume}
}
\value{
Data frame with strategy results
}
\description{
Implements CPPI strategy for commodity price risk management
}

