% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy_classes.R
\docType{class}
\name{DPPI-class}
\alias{DPPI-class}
\title{An S4 class for the DPPI hedging strategy}
\description{
An S4 class for the DPPI hedging strategy
}
\section{Slots}{

\describe{
\item{\code{TargetPercent}}{A percentage of first trading day's market price used to set target price (cap or floor)}

\item{\code{RiskFactor}}{The risk factor (cushion) used in the DPPI model}
}}

