% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DppiStrat.R
\name{DppiStrat}
\alias{DppiStrat}
\title{Dynamic Proportion Portfolio Insurance (DPPI)}
\usage{
DppiStrat(Q, P, tper = 1.15, RF = 10, tcost = 0, min = Q, int = TRUE)
}
\arguments{
\item{Q}{numeric value for quantity to be hedged}

\item{P}{numeric futures price vector}

\item{tper}{numeric target price markup}

\item{RF}{numeric risk factor vector}

\item{tcost}{numeric transaction costs pr unit}

\item{min}{numeric minimum quantity to be hedged at delivery}

\item{int}{TRUE/ FALSE integer restriction on tradable volume}
}
\value{
Data frame with strategy results
}
\description{
Implements DPPI strategy for price risk management
}

