% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etrm-package.R
\docType{package}
\name{etrm-package}
\alias{etrm-package}
\title{etrm: Energy Trading and Risk Management}
\description{
Tools for energy market risk management (forward curves and trading strategies)
}
\references{
F. E. Benth, S. Koekkebakker, and F. Ollmar.  Extracting and applying smooth forward curves from average-based commodity contracts with seasonal variation.The Journal of Derivatives, 15(1):52–66,2007b. https://doi.org/10.3905/jod.2007.694791

F. E. Benth,  J. S. Benth,  and S. Koekebakker. Stochastic modelling of electricity and related markets,volume 11. World Scientific, 2008. https://doi.org/10.1142/6811

F. Black. The pricing of commodity contracts.Journal of financial economics, 3(1):167–179, 1976. https://doi.org/10.1016/0304-405X(76)90024-6

T. Bjork. Arbitrage Theory in Continuous Time.  Oxford University Press, 3 edition, 2009.  https://EconPapers.repec.org/RePEc:oxp:obooks:9780199574742

F. Black and R. W. Jones. Simplifying portfolio insurance. The Journal of Portfolio Management, 14(1):48–51, 1987. https://doi.org/10.3905/jpm.1987.409131

H. E. Leland. Who should buy portfolio insurance? The Journal of Finance, 35(2):581–594, 1980. http://www.jstor.org/stable/2327419
}
\author{
Anders D. Sleire <sleire@gmail.com>
}
