% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msfc_class.R
\docType{class}
\name{MSFC-class}
\alias{MSFC-class}
\title{An S4 class for the Maximum Smoothness Forward Curve (MSFC) in hedgeR}
\description{
An S4 class for the Maximum Smoothness Forward Curve (MSFC) in hedgeR
}

