% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msfc.R
\name{msfc}
\alias{msfc}
\title{Maximum Smoothness Forward Curve (MSFC)}
\usage{
msfc(tdate, include, sdate, edate, f)
}
\arguments{
\item{tdate}{trading date}

\item{include}{logical vector to determine if contracts should be included in calculation}

\item{sdate}{date vector with contract delivery start dates}

\item{edate}{date vector with contract delivery end dates}

\item{f}{numeric vector with futures contract prices}
}
\value{
instance of the MSFC class
}
\description{
Creates a smooth forward curve from futures prices for a flow delivery
}

