% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obpi_strategy.R
\name{obpi}
\alias{obpi}
\title{Option Based Portfolio Insurance (OBPI)}
\usage{
obpi(q, tdate, f, k = f[1], vol, r = 0, tdays = 250, daysleft,
  tcost = 0, int = TRUE)
}
\arguments{
\item{q}{numeric value for quantity to be hedged, either positive (net buyer) or negative (net seller)}

\item{tdate}{date vector with trading days}

\item{f}{numeric futures price vector}

\item{k}{numeric value for option strike price}

\item{vol}{value for volatility}

\item{r}{value for interest rate}

\item{tdays}{integer assumed number of trading days per year}

\item{daysleft}{integer with days left to option expiry}

\item{tcost}{numeric transaction costs pr unit}

\item{int}{TRUE/ FALSE integer restriction on tradable volume}

\item{min}{numeric minimum quantity to be hedged at delivery}
}
\value{
instance of the OBPI class
}
\description{
Implements OBPI strategy for commodity price risk management
}

