% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObpiStrat.R
\name{ObpiStrat}
\alias{ObpiStrat}
\title{Option Based Portfolio Insurance (OBPI)}
\usage{
ObpiStrat(Q, P, K = P[1], vol, R = 0, tdays = 250, tcost = 0, min = Q,
  int = TRUE)
}
\arguments{
\item{Q}{numeric value for quantity to be hedged}

\item{P}{numeric futures price vector}

\item{K}{numeric strike price}

\item{vol}{numeric volatility}

\item{R}{numeric interest rate}

\item{tdays}{integer assumed number of trading days per year}

\item{tcost}{numeric transaction costs pr unit}

\item{min}{numeric minimum quantity to be hedged at delivery}

\item{int}{TRUE/ FALSE integer restriction on tradable volume}
}
\value{
Data frame with strategy results
}
\description{
Implements OBPI strategy for commodity price risk management
}

