% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{powcal}
\alias{powcal}
\title{Historical daily closing prices for 11 calendar year power futures contracts}
\format{
A data frame with 3253 rows and 12 columns:
\describe{
  \item{Date}{the trading date}
  \item{CAL-06}{the closing price for the 2006 futures contract}
  \item{CAL-07}{the closing price for the 2007 futures contract}
  \item{CAL-08}{the closing price for the 2008 futures contract}
  \item{CAL-09}{the closing price for the 2009 futures contract}
  \item{CAL-10}{the closing price for the 2010 futures contract}
  \item{CAL-11}{the closing price for the 2011 futures contract}
  \item{CAL-12}{the closing price for the 2012 futures contract}
  \item{CAL-13}{the closing price for the 2013 futures contract}
  \item{CAL-14}{the closing price for the 2014 futures contract}
  \item{CAL-15}{the closing price for the 2015 futures contract}
  \item{CAL-16}{the closing price for the 2016 futures contract}
}
}
\usage{
powcal
}
\description{
A synthetic dataset containing the closing prices and other attributes of 11
 power futures contracts for calendar year delivery for 2006 - 2016.
}
\keyword{datasets}
