% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{powfutures130513}
\alias{powfutures130513}
\title{Closing prices for power futures contracts at trading date 2013-05-13}
\format{
A data frame with 38 rows and 5 columns:
\describe{
  \item{Include}{boolean variable to determine if contract should be included in forward curve calculation}
  \item{Contract}{the name of the futures contract}
  \item{Start}{delivery start date for the futures contract}
  \item{End}{delivery start date for the futures contract}
  \item{Closing}{the futures contract closing price}
}
}
\usage{
powfutures130513
}
\description{
A synthetic dataset containing the closing prices and other attributes of 38
 power futures contracts.
}
\keyword{datasets}
