% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{powfutures140909}
\alias{powfutures140909}
\title{Closing prices for power futures contracts at trading date 2014-09-09}
\format{
A data frame with 36 rows and 5 columns:
\describe{
  \item{Include}{boolean variable to determine if contract should be included in forward curve calculation}
  \item{Contract}{the name of the futures contract}
  \item{Start}{delivery start date for the futures contract}
  \item{End}{delivery start date for the futures contract}
  \item{Closing}{the futures contract closing price}
}
}
\usage{
powfutures140909
}
\description{
A synthetic dataset containing the closing prices and other attributes of 36
 power futures contracts. The variables are as follows:
}
\keyword{datasets}
