% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{powpriors130513}
\alias{powpriors130513}
\title{Example priors at trading date 2015-05-13}
\format{
A data frame with 3885 rows and 3 columns:
\describe{
  \item{Date}{vector of dates ranging from 2013-05-13 to final end date of contracts in powfutures130513}
  \item{trig.prior}{a simple smooth trigonometric prior describing power price seasonality}
  \item{mod.prior}{a trigonometric prior adjusted for typical calendar effects}
}
}
\usage{
powpriors130513
}
\description{
An example of two simple priors for forward market price to be used with powfutures130513
}
\keyword{datasets}
