% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SfpiStrat.R
\name{SfpiStrat}
\alias{SfpiStrat}
\title{Sigmoid Function Portfolio Insurance (SFPI)}
\usage{
SfpiStrat(Q, P, a, tcost = 0, min = Q, int = TRUE)
}
\arguments{
\item{Q}{numeric value for quantity to be hedged}

\item{P}{numeric futures price vector}

\item{a}{numeric risk appetite parameter}

\item{tcost}{numeric transaction costs pr unit}

\item{min}{numeric minimum quantity to be hedged at delivery}

\item{int}{TRUE/ FALSE integer restriction on tradable volume}
}
\value{
Data frame with strategy results
}
\description{
Implements SFPI strategy for commodity price risk management
}

