% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msfc_class.R
\name{show,MSFC-method}
\alias{show,MSFC-method}
\title{S4 method for the show generic}
\usage{
\S4method{show}{MSFC}(object)
}
\arguments{
\item{object}{instance of the MSFC class}
}
\description{
S4 method for the show generic
}
