% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shpi_strategy.R
\name{shpi}
\alias{shpi}
\title{Step Hedge Portfolio Insurance (SHPI)}
\usage{
shpi(q, tdate, f, daysleft, tper = 0.1, tcost = 0, int = TRUE)
}
\arguments{
\item{q}{numeric value for quantity to be hedged, either positive (net buyer) or negative (net seller)}

\item{tdate}{date vector with trading days}

\item{f}{numeric futures price vector}

\item{daysleft}{integer with days left to contract expiry}

\item{tper}{numeric target price markup/down to the price on the first trading day}

\item{tcost}{numeric transaction costs pr unit}

\item{int}{TRUE/FALSE integer restriction on tradable volume}
}
\value{
instance of the SHPI class
}
\description{
Implements SHPI strategy for commodity price risk management
}
