% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShpiStrat.R
\name{ShpiStrat}
\alias{ShpiStrat}
\title{Step Hedge Portfolio Insurance (SHPI)}
\usage{
ShpiStrat(Q, P, tcost = 0, int = TRUE)
}
\arguments{
\item{Q}{numeric value for quantity to be hedged}

\item{P}{numeric futures price vector}

\item{tcost}{numeric transaction costs pr unit}

\item{int}{TRUE/ FALSE integer restriction on tradable volume}
}
\value{
Data frame with strategy results
}
\description{
Implements SHPI strategy for price risk management
}

