% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VbpiStrat.R
\name{VbpiStrat}
\alias{VbpiStrat}
\title{VaR Based Portfolio Insurance (VBPI)}
\usage{
VbpiStrat(Q, P, TP, conf = 0.95, tcost = 0, min = Q, int = TRUE)
}
\arguments{
\item{Q}{numeric value for quantity to be hedged}

\item{P}{numeric futures price vector}

\item{TP}{numeric target price to protect}

\item{conf}{numeric confidence level}

\item{tcost}{numeric transaction costs pr unit}

\item{min}{numeric minimum quantity to be hedged at delivery}

\item{int}{TRUE/ FALSE integer restriction on tradable volume}
}
\value{
Data frame with strategy results
}
\description{
Implements VBPI strategy for commodity price risk management
}

