% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy_classes.R
\docType{class}
\name{GenericStrat-class}
\alias{GenericStrat-class}
\title{An S4 VIRTUAL parent class for the hedging strategy classes in etrm}
\description{
An S4 VIRTUAL parent class for the hedging strategy classes in etrm
}
