% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy_classes.R
\name{plot,GenericStrat-method}
\alias{plot,GenericStrat-method}
\title{S4 method for the plot generic}
\usage{
\S4method{plot}{GenericStrat}(
  x,
  y = NULL,
  title = "Strategy plot",
  xlab = "",
  ylab.1 = "Price",
  ylab.2 = "Hedge \%",
  pcols = c("#F8766D", "steelblue3", "gray60", "gray80"),
  legend = "bottom",
  ...
)
}
\description{
S4 method for the plot generic
}
