% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msfc_class.R
\name{plot,MSFC-method}
\alias{plot,MSFC-method}
\title{S4 method for the plot generic}
\usage{
\S4method{plot}{MSFC}(x, y = NULL, title = "", xlab = "", ylab = "Price", legend = "top", ...)
}
\description{
S4 method for the plot generic
}
