% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msfc_class.R
\name{plot,MSFC-method}
\alias{plot,MSFC-method}
\title{S4 method for the plot generic for class "MSFC"}
\usage{
\S4method{plot}{MSFC}(
  x,
  y = NULL,
  plot.prior = FALSE,
  title = "",
  xlab = "",
  ylab = "Price",
  legend = "right"
)
}
\arguments{
\item{x}{instance of the MSFC class created by the msfc function}

\item{y}{NULL}

\item{plot.prior}{TRUE/FALSE for incuding prior function in plot}

\item{title}{plot title}

\item{xlab}{x-axis title}

\item{ylab}{y-axis title}

\item{legend}{position of legend, as implemented in ggplot2}
}
\value{
a chart with daily values for the forward curve and contracts used in calculation
}
\description{
S4 method for the plot generic for class "MSFC"
}
