% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy_classes.R
\name{show,GenericStrat-method}
\alias{show,GenericStrat-method}
\title{S4 method for the show generic}
\usage{
\S4method{show}{GenericStrat}(object)
}
\arguments{
\item{object}{instance of a strategy class}
}
\description{
S4 method for the show generic
}
