% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy_classes.R
\docType{methods}
\name{summary,GenericStrat-method}
\alias{summary,GenericStrat-method}
\title{S4 method for the summary generic}
\usage{
\S4method{summary}{GenericStrat}(object, ...)
}
\description{
S4 method for the summary generic
}

